+++ slash/core/fixtures/fixture_base.py
@@ -1,7 +1,7 @@
 class FixtureBase(object):
 
     info = None
-    fixture_kwargs = None
+    keyword_arguments = None
     parametrization_ids = None
     names = None
 
@@ -15,8 +15,8 @@ class FixtureBase(object):
         return None  # pragma: no cover
 
     def resolve(self, store):
-        if self.fixture_kwargs is None:
-            self.fixture_kwargs = self._resolve(store)
+        if self.keyword_arguments is None:
+            self.keyword_arguments = self._resolve(store)
 
     def _resolve(self, store):
         raise NotImplementedError()  # pragma: no cover