+++ doc/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 
 * :feature:`524` ``slash list``, ``slash list-config`` and ``slash list-plugins`` now supports ``--force-color``/``--no-color`` flags. The default changed from colored to colored only for tty.
 * :bug:`516 major` Fire test_interrupt earlier and properly mark session as interrupted when a test is interrupted
+* :feature:`513` Add deep parametrization info (including nested fixtures) to the metadata variation info
 * :feature:`512` ``slash list-config`` now receives a path filter for config paths to display
 * :feature:`519` Add ``--no-output`` flag for ``slash list``
 * :feature:`497` Major overhaul of CLI mechanics -- improve help message and usage, as well as cleaner error exits during the session configuration phase